use doubles instead of floats in glColor*

+++ gl.h
@@ -27,7 +27,7 @@
 
 void gl_clear (void);
 bool gl_get_screen_rgb (image_t *image) GNUC_NONNULL GNUC_WARN_UNUSED_RES;
-void gl_color (float r, float g, float b, float a);
+void gl_color (double r, double g, double b, double a);
 void gl_draw_stretched (int gltex);
 void gl_draw_texture (int gltex,
                       double centerx,
